Module Bap_powerpc_target

Declares some of the PowerPC targets.

More targets could be added without modifying this module. It is advised to inherit from one of the targets defined here to include the newly defined target into the powerpc family, though it is not strictly required.

val parent : Bap_core_theory.Theory.Target.t

parent it the parent of all PowerPC.

No properties are set.

The PowerPC 32 family

val powerpc32bi : Bap_core_theory.Theory.Target.t

The bi-endian PowerPC32 (the parent)

val powerpc32eb : Bap_core_theory.Theory.Target.t

The bi-endian PowerPC32 (the parent)

The big endian PowerPC32

val powerpc32le : Bap_core_theory.Theory.Target.t

The big endian PowerPC32

The little endian PowerPC32

The PowerPC 64 family

val powerpc64bi : Bap_core_theory.Theory.Target.t

The bi-endian PowerPC64 (the parent)

val powerpc64eb : Bap_core_theory.Theory.Target.t

The bi-endian PowerPC64 (the parent)

The big endian PowerPC64

val powerpc64le : Bap_core_theory.Theory.Target.t

The big endian PowerPC64

The little endian PowerPC64

The PowerPC encodings

val llvm_powerpc32 : Bap_core_theory.Theory.language

val llvm_powerpc64 : Bap_core_theory.Theory.language

val load : ?backend:string -> unit -> unit

load () loads the knowledge base rules for the PowerPC targets.

This includes parsing the loader output and enabling backward compatibility with the old Arch.t representation.